pyproject.toml
[tool.poetry]
name = "rpgems"
version = "0.1.0"
description = "A system to create and play an RPG game"
authors = ["Marco Gomes", "Rafael Ribeiro", "Teles"]
readme = "README.md"
# package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
pydantic-settings = "^2.5.2"
alembic = "^1.13.3"
sqlalchemy = "^2.0.36"
asyncpg = "^0.30.0"
dependency-injector = "^4.42.0"
psycopg2-binary = "^2.9.10"

[tool.poetry.group.dev.dependencies]
httpx = "^0.27.2"
coverage = "^7.6.4"
pylint = "^3.3.1"
pytest = "^8.3.3"
mypy = "^1.11.2"
pytest-asyncio = "^0.24.0"
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
